refactor reporting portfolio snapshot route
